Move qlog tracer into its own file
